Proses pemindahan anggota pokja dari cek personil ke data tender telah berlangsung
